structural refactoring of the project [partial]
