+++ core/src/services/video/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod decoder;
+pub(crate) mod service;
+
+pub use service::{VideoService, Video};
+pub use decoder::{VideoDump, VideoFrame, VideoThumb};
