--- src/core/mod.rs
@@ -1,3 +1,0 @@
-pub mod decoder;
-pub mod search;
-pub mod service;
